gen_query/pyproject.toml
[project]
name = "gen-query"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["openai", "python-dotenv"]
